backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.200
opencv-python==4.8.0.76
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-image==0.21.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.5.0
werkzeug==2.3.7
matplotlib==3.7.2
seaborn==0.12.2